api/pyproject.toml
[project]
name = "allumfurniture"
version = "0.1.0"
description = ""
authors = [
    {name = "Semon Horochih",email = "HorSemAlex@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
package-mode = false

dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg (>=3.2.9,<4.0.0)",
    "psycopg-binary (>=3.2.9,<4.0.0)",
]

[tool.poetry.group.test.dependencies]
flake8 = "^7.3.0"
pep8-naming = "^0.15.1"
flake8-pyproject = "^1.2.3"

[tool.flake8]
max-line-length = 120
ignore = [
    # W503: line break before binary operator
    "W503",
]
exclude = [
    ".git",
    "__pycache__",
    "conf",
    "venv",
    ".venv",
    ".vscode",
    ".idea",
    "versions"
]
per-file-ignores = [
    "migrations/env.py:F401",
    "migrations/versions/*:W291",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"